Idea_Generation-for-medical-reasearch/Idea_Generation/requirements.txt
dash
dash-bootstrap-components
python-dotenv
fastapi
uvicorn                     